IGNITE-16566 ClusterNode class to interface and implementation
