+++ modules/core/src/main/java/org/apache/ignite/lang/IgniteExceptionMapper.java
