+++ modules/core/src/main/java/org/apache/ignite/lang/IgniteExceptionMapperUtil.java
